requirements.txt
asgiref==3.6.0
Django==4.1.5
sqlparse==0.4.3
django-crispy-forms==1.14.0
crispy-bootstrap5==0.7
django-allauth==0.52.0
Pillow==9.4.0
